server/requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0